docker/requirements.txt
slack-bolt==1.18.0
google-cloud-bigquery==3.13.0
